Fixed low performance

+++ app/src/main/java/com/example/showplaceproject/core/Utils.kt
@@ -62,7 +62,7 @@ fun Frame.screenCenter(view: View): Vector3 {
     return Vector3(view.width / 2f, view.height / 2f, 0f)
 }
 
-fun formatTime(milliseconds: Int): String {
+fun formatTime(milliseconds: Long): String {
     val minutes = (milliseconds / 1000) / 60
     val seconds = (milliseconds / 1000) % 60
     return "$minutes:$seconds"
